Moving table generation into results_processing
